Add navigatino context with a context processor
